requirements.txt
transformers==4.11.3
torch==2.2.0
tensorboard==2.12.0
pytorch-crf==0.7.2
wandb==0.12.1
torchvision==0.17.0
scikit-learn==1.3.2
seqeval==1.2.2